// src/app/api/contribuyente/models/domicilio-buscar-response.ts
/* tslint:disable */
/* eslint-disable */
export interface DomicilioBuscarResponse {
  contribuyenteNumero?: number;
  domicilioContribuyenteNumero?: number;
  domicilioDsc?: string;
  estadoDsc?: string;
  municipalidadId?: number;
  rowNum?: number;
  tipoDomicilioDsc?: string;
  tipoPredioDsc?: string;
}